requirements.txt
anthropic>=0.40.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
yfinance>=0.2.36
ta>=0.11.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
aiohttp>=3.9.0
pydantic>=2.5.0
rich>=13.7.0
